move autobahn example to rama crate
