+++ rama-ws/autobahn/client.sh
@@ -31,5 +31,5 @@ docker run -d --rm \
     wstest -m fuzzingserver -s 'autobahn/fuzzingserver.json'
 
 sleep 5
-cargo run --release --example autobahn-client
+cargo run --release -p rama --example autobahn_client --features=http-full,ws
 test_diff
